Add test parsing fbautostart control file
