+++ src/lib.rs
@@ -19,4 +19,4 @@ pub use self::changelog::{Changelog, ChangelogEntry};
 pub use self::control::{
     ControlEntry, ControlFile, ControlParagraph, ControlValue,
 };
-pub use self::version::Version;
+pub use self::version::{Version, VersionElement, VersionPart};
