+++ tests/data/control-fbautostart
@@ -1,0 +1,19 @@
+Source: fbautostart
+Section: misc
+Priority: optional
+Maintainer: Paul Tagliamonte <paultag@ubuntu.com>
+Build-Depends: debhelper (>= 9)
+Standards-Version: 3.9.3
+Homepage: https://launchpad.net/fbautostart
+Vcs-Git: git://git.debian.org/collab-maint/fbautostart.git
+Vcs-Browser: http://git.debian.org/?p=collab-maint/fbautostart.git
+
+Package: fbautostart
+Architecture: any
+Depends: ${shlibs:Depends}, ${misc:Depends}
+Description: XDG compliant autostarting app for Fluxbox
+ The fbautostart app was designed to have little to no overhead, while
+ still maintaining the needed functionality of launching applications
+ according to the XDG spec.
+ .
+ This package contains support for GNOME and KDE.
